modal, datepicker for frontend crud

+++ Application/new_application/advanced/frontend/controllers/DirectiveController.php
@@ -125,9 +125,13 @@ class DirectiveController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can('delete-directive')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException;
+        }
     }
 
     /**
